core_nagios: split up into modules
